% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{Client}
\alias{Client}
\title{Client}
\usage{
Client(apiKey)
}
\arguments{
\item{apiKey}{A character string containing the API key.}
}
\value{
Returns the status of the local API key registration.
}
\description{
Sets up client access.
}
\examples{
\dontrun{
Client("56gIvzm4dj1vJpNUlv3RJ2CeMd47JETG3bcf5zLS")
}
}
\author{
Riasat Ali Istiaque, \email{riasat.istiaque@pathlit.io}
}
