% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_info}
\alias{get_info}
\title{Get Information}
\usage{
get_info()
}
\value{
Returns a message indicating PathLit Engine version number.
}
\description{
Obtains the current version of the PathLit Engine. See
\url{https://www.pathlit.io/docs/api/info/} for more information.
}
\examples{
\dontrun{
get_info()
}
}
\author{
Riasat Ali Istiaque, \email{riasat.istiaque@pathlit.io}
}
