% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paths.R
\name{get_paths}
\alias{get_paths}
\title{Get Paths}
\usage{
get_paths(tickers)
}
\arguments{
\item{tickers}{A character vector of instrument tickers.}
}
\value{
Returns a time series of dollar movements across the strategies for
the tickers provided.
}
\description{
Computes the dollar returns of a 100,000 USD portfolio
according to the weights computed by \link[pathlit]{get_weights}. See
\url{https://www.pathlit.io/docs/api/paths/} for more information.
}
\examples{
\dontrun{
get_paths(c("AAPL", "HOG", "KO"))
}
}
\author{
Riasat Ali Istiaque, \email{riasat.istiaque@pathlit.io}
}
