% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries.R
\name{get_timeseries}
\alias{get_timeseries}
\title{Get Time Series}
\usage{
get_timeseries(tickers)
}
\arguments{
\item{tickers}{A character vector of instrument tickers.}
}
\value{
Returns a time series of the tickers provided.
}
\description{
Retrieves the time series of the tickers provided. The
comprehensive list of the tickers can be retrieved via
\link[pathlit]{get_tickers}.
See \url{https://www.pathlit.io/docs/api/paths/} for more information.
}
\examples{
\dontrun{
get_timeseries(c("AAPL", "MSFT"))
}
}
\author{
Riasat Ali Istiaque, \email{riasat.istiaque@pathlit.io}
}
