% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_variables.R
\docType{data}
\name{PROD_HOST}
\alias{PROD_HOST}
\title{Global Variables}
\format{
An object of class \code{character} of length 1.
}
\usage{
PROD_HOST
}
\description{
Contains global variables used throughout the package.
}
\author{
Riasat Ali Istiaque, \email{riasat.istiaque@pathlit.io}
}
\keyword{datasets}
